Rejig tests to use TestCaseSource
